+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix the allocator initialization
 
 ### Changed
-- Increase the default AXI width
+- Increase the default AXI width to 512 for MemPool and TeraPool
 - Register all control signals at hierarchy boundaries
 - Upgrade to LLVM 14